@@ -1,13 +1,15 @@
-module.exports = [
+import type { ModuleOptions } from 'webpack';
+
+export const rules: Required<ModuleOptions>['rules'] = [
   // Add support for native node modules
   {
     // We're specifying native_modules in the test because the asset relocator loader generates a
     // "fake" .node file which is really a cjs file.
-    test: /native_modules\/.+\.node$/,
+    test: /native_modules[/\\].+\.node$/,
     use: 'node-loader',
   },
   {
-    test: /\.(m?js|node)$/,
+    test: /[/\\]node_modules[/\\].+\.(m?js|node)$/,
     parser: { amd: false },
     use: {
       loader: '@vercel/webpack-asset-relocator-loader',
@@ -22,8 +24,8 @@ module.exports = [
     use: {
       loader: 'ts-loader',
       options: {
-        transpileOnly: true
-      }
-    }
+        transpileOnly: true,
+      },
+    },
   },
 ];
